% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{aov.bal.boot}
\alias{aov.bal.boot}
\title{Implementation of a one-way ANOVA balanced bootstrap with a balanced bootstrapped Bonferonni-corrected post-hoc test.}
\usage{
aov.bal.boot(
  formula,
  data = NULL,
  summary.only = T,
  aov.n = 10000,
  aov.center = TRUE,
  aov.power = 0.05,
  do.posthoc = T,
  posthoc.n = 10000,
  posthoc.pairing = FALSE,
  posthoc.function = mean,
  posthoc.power = 0.05
)
}
\arguments{
\item{formula}{A formula in the linear model sense. This is coded *only* for one-way ANOVA, so your formula should be y ~ x, where y is the response variable and x is the grouping variable.}

\item{data}{(optional) data frame that your formula variables belong to. Only needed if you don't input the names of global vectors.}

\item{aov.n}{The number of bootstrap replicates to perform for the ANOVA.}

\item{aov.center}{Should the group means be subtracted when forming the null distribution for the ANOVA? This eliminates odd behaviors (multi-modal distributions) when you have groups with extremely disparate means.}

\item{aov.power}{What is the level of statistical significance for the anova?}

\item{do.posthoc}{Boolean, whether or not to perform the pairwise comparison.}

\item{posthoc.n}{The number of bootstrap replicates to perform for the posthoc testing.}

\item{posthoc.pairing}{Are the entries in each group paired data?}

\item{posthoc.function}{Summary statistic to estimate.}

\item{posthoc.power}{What is the level of statistical significance for the pairwise bootstrap? A Bonferroni procedure adjusts this for each test.}

\item{summary.only:}{If TRUE, a single summary data frame is reported. If FALSE, a three-component list is reported containing these elements:
\itemize{
\item tidy, the aforementioned single summary data.frame
\item aov, the bootstrapped distribution of F
\item ph, the bootstrapped distributions of differences between each group
}}
}
\description{
The one-way ANOVA bootstrap approach here is, broadly, based on David C. Howell's website (https://www.uvm.edu/~statdhtx/StatPages/Resampling/Resampling.html)
whose primary source material is this book: Efron, B. & Tibshirani, R. J. (1993). An introduction to the bootstrap. London: Chapman & Hall.
The general approach of constructing this null distribution is mentioned in that book and is used elsewhere in the peer-reviewed literature.
Briefly, the concept of this test is that we calculate an observed F ratio for the data. Rather than reference that F ratio to a look-up table,
we construct our own look-up table based on the characteristics of our data. To do this, we resample the data in an ungrouped fashion, slice the samples
into bootstrap replicates the size of the original data set and then assign the original group identities to the now shuffled data. This is, essentially,
what our null hypothesis is: what if each group actually came from a single population? Then if we shuffle the group identities randomly and perform the model,
then the bootstrapped F ratio will have the same variance between the groups as the original grouping. The hypothesis test is, then, counting the proportion
of bootstrapped F ratios that are *more* extreme than the observed. A p = 0.05 would indicate that 5% of the bootstrapped F ratios are more extreme than the
original data structure. A potential downside or confusion is that, in the case of strong between-group differences, once shuffled it is entirely possible
that it is *impossible* to have an F ratio more extreme than observed, which would result in a p-value of exactly zero. This doesn't happen with traditional
lookup tables. However, the actual p-value of all tests is not particularly informative - we just want the result of whether or not the p-value is smaller
than a pre-stated level of statistical power, which will typically be p = 0.05.
}
\details{
A pairwise balanced bootstrap is performed as a post-hoc test for each unique pairing of the levels of the input 'x' variable of your formula.
A Bonferroni procedure is baked in that expands the bootstrap confidence intervals based on the number of comparisons.
The 'bal.boot' function is used to perform each bootstrap.

The 'balanced' notion used here is taken from "Efficient bootstrap simulation" by Davison et al. 1986 Biometrika, Volume 73, Issue 3, December 1986, Pages 555–566, 
https://doi.org/10.1093/biomet/73.3.555
}
\section{Output}{

This function returns a named list as the result:
\itemize{
\item tidy.data - Summary statistics of the bootstrapped distribution
\item data - The complete bootstrap distribution(s) summarized at the replicate level using stat.function
\item parameters - The input parameters of the function
\item stat.function - The actual function used as the sampling statistic
\item input - The input data
}
}

