% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{bal.boot}
\alias{bal.boot}
\title{Implementation of the balanced bootstrap for one or two samples}
\usage{
bal.boot(
  a,
  b = NA,
  asd = NA,
  bsd = NA,
  n = 10000,
  ci.width = 95,
  null = 0,
  stat.function = mean,
  paired = F
)
}
\arguments{
\item{a}{First dataset, a numerical vector.}

\item{b}{Second dataset, a numerical vector. If paired=TRUE, then length must be equal to a.}

\item{asd}{(optional) First dataset's error as 1 standard deviation. If used, this must either be a single number (if error is uniform) or a vector with length equal to a.}

\item{bsd}{(optional) Second dataset's error as 1 standard deviation. If used, this must either be a single number (if error is uniform) or a vector with length equal to a.}

\item{n}{Number of bootstrap replicates to perform.}

\item{ci.width}{Width of the confidence interval to use for hypothesis testing, a single numeric value between 1 and 100}

\item{null}{Value representing the null hypothesis, default is 0.}

\item{stat.function}{Sampling statistic to use. Can use any function that takes a single vector and reports a single numerical value as its result. Default is the mean.}

\item{paired}{(optional) Boolean to indicate if the data are paired. Only relevant for two-sample situations.}
}
\description{
This function performs a balanced bootstrap on one or two samples If only one sample is provided, its bootstrapped distribution is compared against 
a defined value (null) to test for significance. If two samples are provided, then the first sample's summary statistic is subtracted from the second
sample's sampling statistic and the difference is compared against the defined value (null) to test for signifiance.
}
\details{
By default, this is a simple balanced bootstrap where the sample is replicates n times, shuffled, and then subset into groups equal to the original
sample size. However, if errors (asd or bsd) for either group are provided, then each bootstrap replicate for a value with an associated error is sampled from a 
normal distribution whose mean is the observed value and standard deviation is the user-submitted error. Thus, if errors are provided, this is more 
of a balanced Monte Carlo and carries an additional assumption that the true distribution surrounding individual values is normal.

The 'balanced' notion used here is taken from "Efficient bootstrap simulation" by Davison et al. 1986 Biometrika, Volume 73, Issue 3, December 1986, Pages 555–566, 
https://doi.org/10.1093/biomet/73.3.555

The 'difference' tested is b minus a, thus the directionality is such that positive differences indicate b > a and negative differences indicate b < a.
}
\section{Output}{

This function returns a named list as the result:
\itemize{
\item tidy.data - Summary statistics of the bootstrapped distribution
\item data - The complete bootstrap distribution(s) summarized at the replicate level using stat.function
\item parameters - The input parameters of the function
\item stat.function - The actual function used as the sampling statistic
\item input - The input data
}
}

