% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{bin_round}
\alias{bin_round}
\title{Bins a vector via rounding to the nearest bin.}
\usage{
bin_round(x, bin)
}
\arguments{
\item{x}{A numeric vector.}

\item{bin}{Bin to round towards.}
}
\description{
Bins a vector via rounding to the nearest bin.
}
