% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{cbind.na}
\alias{cbind.na}
\title{cbind with NAs... written by someone on stackoverflow.}
\usage{
\method{cbind}{na}(..., deparse.level = 1)
}
\description{
cbind with NAs... written by someone on stackoverflow.
}
