% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{geom_lm}
\alias{geom_lm}
\title{A simple linear model geom.}
\usage{
geom_lm(..., formula = y ~ x, method = lm, se = F)
}
\description{
A simplified edit of geom_smooth that just draws a y~x OLS regression line.
}
