% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{geom_lmboot}
\alias{geom_lmboot}
\title{Do a balanced bootstrap on a linear model.}
\usage{
geom_lmboot(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  show.legend = NA,
  position = "identity",
  ...,
  n = 10000,
  show.fit = T,
  show.label = T,
  linesize = 1,
  location = "tl",
  method = "ols",
  m = T,
  m.95 = T,
  int = T,
  int.95 = T,
  r2 = T,
  r2.95 = F,
  p = T,
  fontfamily = "serif",
  ci.width = 95,
  show.pred.band = F,
  fill = "blue",
  pred.alpha = 0.25,
  pred.steps = 25,
  fontsize = 6,
  coef.digits = 2,
  nudge_x = 0,
  nudge_y = 0,
  inherit.aes = T
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}}
or \code{\link[ggplot2]{aes_}}.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer,
as a string.}

\item{n}{Bootstrap N, defaults to 10000}

\item{show.fit}{Should the regression fit line be shown? Boolean, defaults to TRUE.}

\item{show.label}{Should the regression fit summary be shown? Boolean, defaults to TRUE.}

\item{linesize}{The size of the fit line (in points).}

\item{location}{Location to print fit results. Defaults to `tl` (top-left), but may be `tr`, `bl`, or `br`.}

\item{method}{Regression model of the slope. Defaults to `ols` for Ordinary Least Squares, but may also be `rma` to perform a Reduced Major Axis (i.e., Standard Major Axis - not Ranged)}

\item{m}{Should the slope coefficient be shown? Boolean, defaults to TRUE.}

\item{m.95}{Should the slope's confidence interval be shown? Boolean, defaults to TRUE.}

\item{int}{Should the intercept be shown? Boolean, defaults to TRUE.}

\item{int.95}{Should the intercept's confidence interval be shown? Boolean, defaults to TRUE.}

\item{r2}{Should Pearson's r-squared be shown? Boolean, defaults to TRUE.}

\item{r2.95}{Should the r-squared confidence interval be shown? Boolean, defaults to FALSE. This can be misleading for poor fits because the calculated r value may be either negative or positive and is then squared. Use with caution.}

\item{p}{Should the slope's p-value be printed? This only prints if the hypothetical 'p' is above or below the significance level chosen. I.e., does the slope's confidence interval overlap with zero?}

\item{fontfamily}{Font family as a string to use for label plotting.}

\item{ci.width}{The width of the confidence interval to use (in percents).}

\item{show.pred.band}{Should the prediction interval be calculated and shown? Boolean. This is computationally intensive because it is also bootstrapped.}

\item{fill}{Color of the prediction interval.}

\item{pred.alpha}{Alpha of the prediction interval.}

\item{pred.steps}{Number of steps to calculate the prediction interval at. Increase for higher resolution at cost of computation time.}

\item{fontsize}{Font size as an integer (in pointS). Used for lable plotting.}

\item{coef.digits}{Number of digits to round to in the fit label.}

\item{nudge_x}{Manual X-axis adjustment of the text. In plotting units (where plotting area ranges from 0 to 1).}

\item{nudge_y}{Manual Y-axis adjustment of the text. In plotting units (where plotting area ranges from 0 to 1).}
}
\description{
This function calculates a linear regression of y versus x. The data
are bootstrapped to generate confidence intervals around the slope and
intercept. Optionally, the r2 value may also be calculated. By default,
the geom prints the text of the results in the upper-left corner of the
plot and also plots the fit line. The bootstrapped prediction interval
around the fit may also be plotted.
}
\section{Aesthetics}{

The geom understands the following aesthetics (required are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item xsd (Vector with length 1 or equal to X. If present, used as the SD term of an rnorm call for each X value during bootstrapping)
\item ysd (Vector with length 1 or equal to Y. If present, used as the SD term of an rnorm call for each Y value during bootstrapping)
}
}

