% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{geom_lmfreq}
\alias{geom_lmfreq}
\title{Perform a linear regression using the lm command and output the results..}
\usage{
geom_lmfreq(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  show.legend = NA,
  position = "identity",
  ...,
  show.fit = T,
  show.label = T,
  linesize = 1,
  location = "tl",
  m = T,
  m.err = T,
  int = T,
  int.err = T,
  r2 = T,
  p = T,
  significance.level = 0.05,
  fontfamily = "serif",
  fontsize = 6,
  coef.digits = 2,
  nudge_x = 0,
  nudge_y = 0,
  inherit.aes = T
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}}
or \code{\link[ggplot2]{aes_}}.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer,
as a string.}

\item{show.fit}{Should the regression fit line be shown? Boolean, defaults to TRUE.}

\item{show.label}{Should the regression fit summary be shown? Boolean, defaults to TRUE.}

\item{linesize}{The size of the fit line (in points).}

\item{location}{Location to print fit results. Defaults to `tl` (top-left), but may be `tr`, `bl`, or `br`.}

\item{m}{Should the slope coefficient be shown? Boolean, defaults to TRUE.}

\item{m.err}{Should the slope's standard error be shown? Boolean, defaults to TRUE.}

\item{int}{Should the intercept be shown? Boolean, defaults to TRUE.}

\item{int.err}{Should the intercept's standard error be shown? Boolean, defaults to TRUE.}

\item{r2}{Should Pearson's r-squared be shown? Boolean, defaults to TRUE.}

\item{p}{Should the slope's significance be shown?}

\item{fontfamily}{Font family as a string to use for label plotting.}

\item{fontsize}{Font size as an integer (in pointS). Used for lable plotting.}

\item{coef.digits}{Number of digits to round to in the fit label.}

\item{nudge_x}{Manual X-axis adjustment of the text. In plotting units (where plotting area ranges from 0 to 1).}

\item{nudge_y}{Manual Y-axis adjustment of the text. In plotting units (where plotting area ranges from 0 to 1).}

\item{signifiance.level}{Significance level to use for hypothesis testing. Default is 0.05}
}
\description{
Perform a linear regression using the lm command and output the results..
}
\section{Aesthetics}{

The geom understands the following aesthetics (required are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
}
}

