% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_lm}
\alias{get_lm}
\title{Take two numerical vectors and output a dataframe containing the OLS slope and intercept}
\usage{
get_lm(x, y)
}
\description{
Take two numerical vectors and output a dataframe containing the OLS slope and intercept
}
