% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{ggsave3}
\alias{ggsave3}
\title{Edit of ggplot save (actually the cowplot implementation ggsave2) that includes some sensible defaults, defaults units to mm,
and uses ImageMagick to trim whitespace from an exported png.}
\usage{
ggsave3(
  filename,
  plot = ggplot2::last_plot(),
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("mm"),
  dpi = 600,
  limitsize = TRUE,
  border = 50,
  ...
)
}
\description{
Intended usage with ImageMagick is to have a set aspect ratio within ggplot (aspect.ratio entry in theme()) and define only
a width (or height) for the saved plot. This will then dynamically trim whitespace for the unspecified dimension.
}
