% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_settings.R
\name{jah_settings}
\alias{jah_settings}
\title{Define a Windows font, set scipen options, and define JAH's default ggplot theme.}
\usage{
jah_settings()
}
\description{
Define a Windows font, set scipen options, and define JAH's default ggplot theme.
}
