% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{left}
\alias{left}
\title{Equivalent of the "left" function in Excel}
\usage{
left(x, n)
}
\arguments{
\item{x}{A text string.}

\item{n}{Number of characters to extract starting from the beginning of the string.}
}
\description{
Equivalent of the "left" function in Excel
}
