% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{lm.bal}
\alias{lm.bal}
\title{Do a balanced bootstrap on a linear model.}
\usage{
lm.bal(
  x,
  y,
  xsd = 0,
  ysd = 0,
  n = 10000,
  ci.width = 95,
  method = c("ols", "rma"),
  pred.band = T,
  pred.steps = 25
)
}
\arguments{
\item{x}{x-axis values, a numeric vector}

\item{y}{y-axis values, a numeric vector}

\item{xsd}{(optional) Either a single value or a vector with length equal to x. Used to generate normal distributions around each x value with SD equal to xsd.}

\item{ysd}{(optional) Either a single value or a vector with length equal to y. Used to generate normal distributions around each y value with SD equal to ysd.}

\item{n}{Number of bootstrap replicates to perform.}

\item{ci.width}{Width of the confidence interval to use for hypothesis testing, a single numeric value between 1 and 100}

\item{method}{regression method. either 'ols" for Ordinary Least Squares or 'rma' for Reduced Major Axis.}

\item{pred.band}{Compute the prediction interval? This is computationally intensive, especially for large datasets.}

\item{pred.steps}{# of steps to calculate the prediction interval at. Increase for higher resolution at cost of computation time.}
}
\description{
This function calculates a linear regression of y versus x. The data
are bootstrapped to generate confidence intervals around the slope and
intercept. Optionally, the r2 value may also be calculated. The bootstrapped
prediction interval around the fit may also be calculated.
}
