% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{mean_ci_dist}
\alias{mean_ci_dist}
\title{Calculate the mean and confidence intervals of a distribution - output designed for stat_summary in ggplot}
\usage{
mean_ci_dist(x, ci = 95)
}
\description{
Calculate the mean and confidence intervals of a distribution - output designed for stat_summary in ggplot
}
