% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_settings.R
\name{mrs_settings}
\alias{mrs_settings}
\title{Define a Windows font, set scipen options, and define MRS's default ggplot theme.}
\usage{
mrs_settings()
}
\description{
Define a Windows font, set scipen options, and define MRS's default ggplot theme.
}
