% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{one.samp.bal.sd}
\alias{one.samp.bal.sd}
\title{Perform a one-sample balanced bootstrap using standard deviation as the sampling statistic}
\usage{
one.samp.bal.sd(a, n = 10000)
}
\arguments{
\item{a}{Input data, a numerical vector.}

\item{n}{Number of bootstrap replicates to perform.}
}
\description{
Perform a one-sample balanced bootstrap using standard deviation as the sampling statistic
}
