% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_functions.R
\name{pca_data}
\alias{pca_data}
\title{Extract PCA data and grouping for further analysis or publication-quality plots.}
\usage{
pca_data(data, cols, groups, scale = T, var_scaling = 5)
}
\description{
Extract PCA data and grouping for further analysis or publication-quality plots.
}
