% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_functions.R
\name{pca_plot}
\alias{pca_plot}
\title{'Clean' looking PCA plotting function.}
\usage{
pca_plot(
  data,
  cols,
  color = NA,
  shape = NA,
  label = NA,
  scale = T,
  var_scaling = 5,
  text_size = 8,
  legend_position = "top",
  font_family = "Times",
  axis_alpha = 0.5,
  geom_type = "text",
  point_size = 4,
  repel_variables = F,
  repel_samples = F,
  point_outline = F
)
}
\description{
'Clean' looking PCA plotting function.
}
