% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{rbind.na}
\alias{rbind.na}
\title{rbind with NAs. Written by someone on stackoverflow.}
\usage{
\method{rbind}{na}(..., deparse.level = 1)
}
\description{
Same functionality as rbind, but with NAs. I honestly don't fully remember the functionality here.
}
