% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{right}
\alias{right}
\title{Equivalent of the "right" function in Excel}
\usage{
right(x, n)
}
\arguments{
\item{x}{A text string.}

\item{n}{Number of characters to extract starting from the end of the string counting backwards.}
}
\description{
Equivalent of the "right" function in Excel
}
