% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{round_df}
\alias{round_df}
\title{Round all numeric entries in a data frame}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{A data.frame}

\item{digits}{Number of digits to round numerical entries to.}
}
\description{
Round all numeric entries in a data frame
}
