% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{squash_axis}
\alias{squash_axis}
\title{A ggplot2 continuous axis transformation used to compress (or expand) a portion of the axis according to a multiplier. 
Specifically written to be used as a scale_continuous 'transform' parameter option.}
\usage{
squash_axis(from, to, factor, base_trans = identity_trans())
}
\arguments{
\item{from}{Axis value to begin the transformation.}

\item{to}{Axis vaoue to end the transformation.}

\item{factor}{Multiplication factor. Values >1 compress and values <1 expand the from/to region of the axis.}

\item{base_trans}{Allows usage with base ggplot2 transform options. These may work oddly when combined so take care.}
}
\description{
A ggplot2 continuous axis transformation used to compress (or expand) a portion of the axis according to a multiplier. 
Specifically written to be used as a scale_continuous 'transform' parameter option.
}
