% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{two.samp.bal}
\alias{two.samp.bal}
\title{Perform a two-sample balanced bootstrap using mean as the sampling statistic, defaults to unpaired}
\usage{
two.samp.bal(a, b, n = 10000, paired = F, dist = F)
}
\description{
Perform a two-sample balanced bootstrap using mean as the sampling statistic, defaults to unpaired
}
