% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{two.samp.bal.sd}
\alias{two.samp.bal.sd}
\title{Perform a two-sample PAIRED balanced bootstrap using standard deviation as the sampling statistic}
\usage{
two.samp.bal.sd(a, b, n = 10000)
}
\description{
Perform a two-sample PAIRED balanced bootstrap using standard deviation as the sampling statistic
}
