% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{one.samp.bal.median}
\alias{one.samp.bal.median}
\title{Perform a one-sample balanced bootstrap using median as the sampling statistic}
\usage{
one.samp.bal.median(a, n = 10000)
}
\description{
Perform a one-sample balanced bootstrap using median as the sampling statistic
}
