% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{one.samp.bal}
\alias{one.samp.bal}
\title{Perform a one-sample balanced bootstrap using mean as the sampling statistic}
\usage{
one.samp.bal(a, n = 10000, incdata = F)
}
\description{
Perform a one-sample balanced bootstrap using mean as the sampling statistic
}
