% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{two.samp.bal.sd}
\alias{two.samp.bal.sd}
\title{Perform a two-sample PAIRED balanced bootstrap using standard deviation as the sampling statistic}
\usage{
two.samp.bal.sd(a, b, n = 10000)
}
\arguments{
\item{a}{First dataset, a numerical vector.}

\item{b}{Second dataset, a numerical vector. Length must be equal to a.}

\item{n}{Number of bootstrap replicates to perform.}
}
\description{
Perform a two-sample PAIRED balanced bootstrap using standard deviation as the sampling statistic
}
